Base framework changes
Renamed the application to Foodtruck.
Create a new page layout and navigation bar.
Added menu page with different categories.
Using the AngularJS UI route for the view change
Using the Spring hibernate jpa.
Renamed the database schema to food_truck
Using ozlazyload for loading the view specific JS and Css and templates.
Using new index1.html file.

+++ foodtruck/src/main/java/com/food/dao/ItemDaoImpl.java
@@ -43,7 +43,7 @@ public class ItemDaoImpl implements ItemDao{
 					  int i = rs.getInt("id");
 					  String title = rs.getString("title");
 					  String link = rs.getString("link");
-					  int price = rs.getInt("price");
+					  Double price = rs.getDouble("price");
 					  Item item = new Item(i,title,link,price);
 
 					  tempItem.add(item);
